+++ src/DAMA/DoctrineTestBundle/DAMADoctrineTestBundle.php
@@ -3,6 +3,7 @@
 namespace DAMA\DoctrineTestBundle;
 
 use DAMA\DoctrineTestBundle\DependencyInjection\DoctrineTestCompilerPass;
+use Symfony\Component\DependencyInjection\Compiler\PassConfig;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -14,6 +15,7 @@ class DAMADoctrineTestBundle extends Bundle
     public function build(ContainerBuilder $container): void
     {
         parent::build($container);
-        $container->addCompilerPass(new DoctrineTestCompilerPass());
+        // lower priority than CacheCompatibilityPass from DoctrineBundle
+        $container->addCompilerPass(new DoctrineTestCompilerPass(), PassConfig::TYPE_BEFORE_OPTIMIZATION, -1);
     }
 }