+++ tests/DAMA/DoctrineTestBundle/DependencyInjection/DoctrineTestCompilerPassTest.php
@@ -4,15 +4,28 @@ namespace Tests\DAMA\DoctrineTestBundle\DependencyInjection;
 
 use DAMA\DoctrineTestBundle\DependencyInjection\DAMADoctrineTestExtension;
 use DAMA\DoctrineTestBundle\DependencyInjection\DoctrineTestCompilerPass;
+use DAMA\DoctrineTestBundle\Doctrine\Cache\Psr6StaticArrayCache;
 use DAMA\DoctrineTestBundle\Doctrine\Cache\StaticArrayCache;
 use Doctrine\Bundle\DoctrineBundle\ConnectionFactory;
+use Doctrine\Common\Cache\ArrayCache;
 use Doctrine\DBAL\Connection;
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\Cache\Adapter\ArrayAdapter;
+use Symfony\Component\DependencyInjection\ChildDefinition;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
 
 class DoctrineTestCompilerPassTest extends TestCase
 {
+    private const CACHE_SERVICE_IDS = [
+        'doctrine.orm.a_metadata_cache',
+        'doctrine.orm.b_metadata_cache',
+        'doctrine.orm.c_metadata_cache',
+        'doctrine.orm.a_query_cache',
+        'doctrine.orm.b_query_cache',
+        'doctrine.orm.c_query_cache',
+    ];
+
     /**
      * @dataProvider processDataProvider
      */
@@ -31,7 +44,7 @@ class DoctrineTestCompilerPassTest extends TestCase
         $containerBuilder->setDefinition('doctrine.dbal.connection_factory', new Definition(ConnectionFactory::class));
 
         if ($expectationCallback !== null) {
-            $expectationCallback($this);
+            $expectationCallback($this, $containerBuilder);
         }
 
         (new DoctrineTestCompilerPass())->process($containerBuilder);
@@ -41,12 +54,14 @@ class DoctrineTestCompilerPassTest extends TestCase
 
     public function processDataProvider(): \Generator
     {
+        $defaultConfig = [
+            'enable_static_connection' => true,
+            'enable_static_meta_data_cache' => true,
+            'enable_static_query_cache' => true,
+        ];
+
         yield 'default config' => [
-            [
-                'enable_static_connection' => true,
-                'enable_static_meta_data_cache' => true,
-                'enable_static_query_cache' => true,
-            ],
+            $defaultConfig,
             function (ContainerBuilder $containerBuilder): void {
                 $this->assertTrue($containerBuilder->hasDefinition('dama.doctrine.dbal.connection_factory'));
                 $this->assertSame(
@@ -54,21 +69,9 @@ class DoctrineTestCompilerPassTest extends TestCase
                     $containerBuilder->getDefinition('dama.doctrine.dbal.connection_factory')->getDecoratedService()[0]
                 );
 
-                $cacheServiceIds = [
-                    'doctrine.orm.a_metadata_cache',
-                    'doctrine.orm.b_metadata_cache',
-                    'doctrine.orm.c_metadata_cache',
-                    'doctrine.orm.a_query_cache',
-                    'doctrine.orm.b_metadata_cache',
-                    'doctrine.orm.c_metadata_cache',
-                ];
-
-                foreach ($cacheServiceIds as $id) {
+                foreach (self::CACHE_SERVICE_IDS as $id) {
                     $this->assertFalse($containerBuilder->hasAlias($id));
-                    $this->assertEquals(
-                        (new Definition(StaticArrayCache::class))->addMethodCall('setNamespace', [sha1($id)]),
-                        $containerBuilder->getDefinition($id)
-                    );
+                    $this->assertFalse($containerBuilder->hasDefinition($id));
                 }
 
                 $this->assertSame([
@@ -130,6 +133,75 @@ class DoctrineTestCompilerPassTest extends TestCase
             function (TestCase $testCase): void {
                 $testCase->expectException(\InvalidArgumentException::class);
                 $testCase->expectExceptionMessage('Unknown doctrine dbal connection name(s): foo, bar.');
+            },
+        ];
+
+        yield 'legacy doctrine/cache ORM services' => [
+            $defaultConfig,
+            function (ContainerBuilder $containerBuilder): void {
+                foreach (self::CACHE_SERVICE_IDS as $id) {
+                    $this->assertFalse($containerBuilder->hasAlias($id));
+                    $this->assertEquals(
+                        (new Definition(StaticArrayCache::class))->addMethodCall('setNamespace', [sha1($id)]),
+                        $containerBuilder->getDefinition($id)
+                    );
+                }
+            },
+            function (self $testCase, ContainerBuilder $containerBuilder): void {
+                foreach (self::CACHE_SERVICE_IDS as $id) {
+                    $containerBuilder->register($id, ArrayCache::class);
+                }
+            },
+        ];
+
+        yield 'psr6 ORM cache services' => [
+            $defaultConfig,
+            function (ContainerBuilder $containerBuilder): void {
+                foreach (self::CACHE_SERVICE_IDS as $id) {
+                    $this->assertFalse($containerBuilder->hasAlias($id));
+                    $this->assertEquals(
+                        (new Definition(Psr6StaticArrayCache::class))->setArgument(0, sha1($id)),
+                        $containerBuilder->getDefinition($id)
+                    );
+                }
+            },
+            function (self $testCase, ContainerBuilder $containerBuilder): void {
+                foreach (self::CACHE_SERVICE_IDS as $id) {
+                    $containerBuilder->register($id, ArrayAdapter::class);
+                }
+            },
+        ];
+
+        yield 'psr6 ORM cache services using child definitions' => [
+            $defaultConfig,
+            function (ContainerBuilder $containerBuilder): void {
+                foreach (self::CACHE_SERVICE_IDS as $id) {
+                    $this->assertFalse($containerBuilder->hasAlias($id));
+                    $this->assertEquals(
+                        (new Definition(Psr6StaticArrayCache::class))->setArgument(0, sha1($id)),
+                        $containerBuilder->getDefinition($id)
+                    );
+                }
+            },
+            function (self $testCase, ContainerBuilder $containerBuilder): void {
+                $parentDefinition = new Definition(ArrayAdapter::class);
+                $containerBuilder->setDefinition('some_cache_parent_definition', $parentDefinition);
+                foreach (self::CACHE_SERVICE_IDS as $id) {
+                    $containerBuilder->setDefinition($id, new ChildDefinition('some_cache_parent_definition'));
+                }
+            },
+        ];
+
+        yield 'invalid ORM cache services' => [
+            $defaultConfig,
+            function (ContainerBuilder $containerBuilder): void {
+            },
+            function (self $testCase, ContainerBuilder $containerBuilder): void {
+                $testCase->expectException(\InvalidArgumentException::class);
+                $testCase->expectExceptionMessage('Unsupported cache class "stdClass" found on service "doctrine.orm.a_metadata_cache"');
+                foreach (self::CACHE_SERVICE_IDS as $id) {
+                    $containerBuilder->register($id, \stdClass::class);
+                }
             },
         ];
     }