+++ CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [x.x.x]
+## [6.6.0]
 ### Changed
 - dropped support for unmaintained Symfony versions 3.4 and 5.1
+- fixed compatibility with DoctrineBundle 2.4+ and ORM 2.9+ by introducing PSR-6 compatible cache
+- added dependency on `symfony/cache`
 
 ## [6.5.0]
 ### Changed